broke files and components apart, added checkmark
