+++ src/components/VersionSelector/ArrowSvg.tsx
@@ -1,0 +1,20 @@
+import * as React from 'react';
+import { ArrowIconProps } from './types';
+
+export const ArrowSvg = ({ className, style }: ArrowIconProps): JSX.Element => (
+  <svg
+    className={className}
+    style={style}
+    xmlns="http://www.w3.org/2000/svg"
+    width="16"
+    height="16"
+    viewBox="0 0 24 24"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="2"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+  >
+    <polyline points="6 9 12 15 18 9" />
+  </svg>
+);
