+++ src/components/VersionSelector/types.ts
@@ -1,11 +1,3 @@
-export interface VersionOption {
-  idx?: number;
-  value: string;
-  title?: string;
-  serverUrl?: string;
-  label?: string;
-}
-
 export interface ActiveVersionData {
   resource_version: string;
   api_version: string;
@@ -14,7 +6,6 @@ export interface VersionSelectorProps {
   resource_versions: string[];
   active: ActiveVersionData;
   root_url: string;
-  onChange?: (option: VersionOption) => void;
   ariaLabel?: string;
   className?: string;
   placeholder?: string;
